+++ Program.cs
@@ -1,0 +1,32 @@
+using Microsoft.EntityFrameworkCore;
+using WebApplication1.Controllers;
+using WebApplication1.Data;
+
+var builder = WebApplication.CreateBuilder(args);
+
+#region Configure MSSQL
+// Add services to the container.
+builder.Services.AddDbContext<CarShopDBContext>(o => o.UseSqlServer(builder.Configuration.GetConnectionString("DefaultConnection")));
+#endregion
+
+builder.Services.AddControllers();
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
