--- WebApplication1/Data/CarShopDBContext.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-using WebApplication1.Models;
-
-namespace WebApplication1.Data
-{
-    public class CarShopDBContext : DbContext
-    {
-        public CarShopDBContext(DbContextOptions options) : base(options) { }
-
-        public DbSet<Car> Cars { get; set; }
-        public DbSet<Orders> Order { get; set; }
-
-
-    }
-}
