--- WebApplication1/Models/Orders.cs
@@ -1,12 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace WebApplication1.Models
-{
-    public class Orders
-    {
-        public int Id { get; set; }
-        public string date { get; set; }
-        public string name { get; set; }
-        public string status { get; set; }
-    }
-}
