--- WebApplication1/WebApplication1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33516.290
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApplication1", "WebApplication1.csproj", "{1828E1D9-F35C-49FF-A427-B905AC63C050}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1828E1D9-F35C-49FF-A427-B905AC63C050}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1828E1D9-F35C-49FF-A427-B905AC63C050}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1828E1D9-F35C-49FF-A427-B905AC63C050}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1828E1D9-F35C-49FF-A427-B905AC63C050}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8DE15FF5-AFB9-4C34-BEF3-7A8EF86AEB81}
-	EndGlobalSection
-EndGlobal
