--- WebApplication1/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "DefaultConnection": "server=COMP-DREAMSGUY;database=DB;TrustServerCertificate=True;Trusted_Connection=True"
-  }
-}
